Moved the validation for the raw cqc downloads
