+++ terraform/pipeline/glue.tf
@@ -752,7 +752,7 @@ module "validate_ascwds_worker_raw_data_job" {
 
 module "validate_locations_api_raw_data_job" {
   source          = "../modules/glue-job"
-  script_dir      = "jobs"
+  script_dir      = "projects/_01_ingest/cqc_api/jobs"
   script_name     = "validate_locations_api_raw_data.py"
   glue_role       = aws_iam_role.sfc_glue_service_iam_role
   resource_bucket = module.pipeline_resources
@@ -767,7 +767,7 @@ module "validate_locations_api_raw_data_job" {
 
 module "validate_providers_api_raw_data_job" {
   source          = "../modules/glue-job"
-  script_dir      = "jobs"
+  script_dir      = "projects/_01_ingest/cqc_api/jobs"
   script_name     = "validate_providers_api_raw_data.py"
   glue_role       = aws_iam_role.sfc_glue_service_iam_role
   resource_bucket = module.pipeline_resources
